+++ ofw/task/updateCheck.php
@@ -1,0 +1,67 @@
+<?php
+class updateCheckTask{
+  public function __toString(){
+    return "updateCheck: Función para comprobar si existen actualizaciones del Framework.";
+  }
+
+  private $repo_url = 'https://raw.githubusercontent.com/igorosabel/Osumi-Framework/master/';
+
+  function doUpdateCheck($repo_version){
+    global $c;
+    $updates = json_decode( file_get_contents($this->repo_url.'ofw/base/updates.json'), true );
+    echo "  ".$updates[$repo_version]['message']."\n";
+    echo "==============================================================================================================\n\n";
+
+    if (count($updates[$repo_version]['deletes'])>0){
+      echo "  Archivos que serán eliminados:\n\n";
+      foreach ($updates[$repo_version]['deletes'] as $delete){
+        $local_delete = $c->getDir('base').$delete;
+        if (file_exists($local_delete)){
+          echo "    ".$local_delete."\n";
+        }
+      }
+      echo "\n";
+    }
+    if (count($updates[$repo_version]['files'])>0){
+      echo "  Archivos actualizados:\n\n";
+      foreach ($updates[$repo_version]['files'] as $file){
+        $local_file = $c->getDir('base').$file;
+        if (file_exists($local_file)){
+          echo "    Actualización: \"".$local_file."\"\n";
+        }
+        else{
+          echo "    Nuevo: \"".$local_file."\"\n";
+        }
+      }
+      echo "\n";
+    }
+    echo "Para proceder a la actualización debes ejecutar el siguiente comando:\n\n";
+    echo "  php ofw.php update\n";
+  }
+
+  public function run(){
+    $current_version = trim( Base::getVersion() );
+    $repo_version = trim( file_get_contents($this->repo_url.'ofw/base/VERSION') );
+
+    echo "Versión instalada: ".$current_version."\n";
+    echo "Versión actual: ".$repo_version."\n";
+
+    $compare = version_compare($current_version, $repo_version);
+
+    switch ($compare){
+      case -1: {
+        echo "La actualización modificará los siguientes archivos:\n";
+        $this->doUpdateCheck($repo_version);
+      }
+      break;
+      case 0: {
+        echo "La versión instalada está actualizada.\n";
+      }
+      break;
+      case 1: {
+        echo "¡¡La versión instalada está MÁS actualizada que la del repositorio!!\n";
+      }
+      break;
+    }
+  }
+}